requirements.txt
# Dependencias principales de Django
Django>=4.2

# Conexión a base de datos (PostgreSQL/Supabase)
dj-database-url
psycopg2-binary
python-dotenv

# Almacenamiento de archivos en Supabase
django-storages[supabase]